Abstraccion de manejo de websocket de roomEchoHandler
